operator: split attachedResources into its own file
(cherry picked from commit 564db32dfd1bfef181a9d077acc19442347c6002)
